"#17-Depuración fallos"

+++ 0013-#Tarea 1-Sprint3/src/com/afundacionfp/JDBCDataProvider.java
@@ -214,15 +214,14 @@ public class JDBCDataProvider implements DataProvider {
                         System.out.println("OK,creado");
                         throw new HttpExceptionCode(201);
                     } else {
-                        System.out.println("Autentificación fallida");
-                        throw new HttpExceptionCode(401);
+
                     }
                 } else {
-                    System.out.println("Not Found");
-                    throw new HttpExceptionCode(404);
+                    System.out.println("Autentificación fallida");
+                    throw new HttpExceptionCode(401);
                 }
             }
-            throw new HttpExceptionCode(418);
+            throw new HttpExceptionCode(404);
         } catch (SQLException throwables) {
             throwables.printStackTrace();
         }
@@ -232,40 +231,37 @@ public class JDBCDataProvider implements DataProvider {
 
 
     @Override
-    public void removeReserve(String reference, String username, String passwordSha) {
+    public void removeReserve(String reference, String username, String passwordSha) throws HttpExceptionCode{
         try {
             // Autenticación
             Connection connection = DriverManager.getConnection("jdbc:mysql://gitlab.afundacionfp.com:3306/mysql?serverTimezone=Europe/Madrid", "developer", "pass");
             Statement statement = connection.createStatement();
             String sql = "SELECT * FROM TablaClientes WHERE usuario = '" + username + "'";
             ResultSet resultSet = statement.executeQuery(sql);
-            while (resultSet.next()) {
+            if (resultSet.next()) {
                 String salt = resultSet.getString(5);
                 String databaseConcatenatedStringSha = resultSet.getString(4);
                 String concatenatedString = salt + passwordSha;
                 String sha1 = sha1FromString(concatenatedString);
+
                 if (databaseConcatenatedStringSha.equals(sha1)) {
                     // Si la autenticación es exitosa, lanzamos una sentencia INSERT como la especificada al final en
                     // https://gitlab.afundacionfp.com/data/acceso-a-datos/-/wikis/Tablas-de-la-base-de-datos-de-Monster-Trucks
                     String sql2 = "DELETE FROM TablaReservas WHERE (idCliente = '" + username + "') AND (refCamion = '" + reference + "')";
                     // Usamos 'executeUpdate' en lugar de 'executeQuery', porque actualizamos la tabla
                     int rowsAffected = statement.executeUpdate(sql2);
-                    if (rowsAffected > 0) {
-                        System.out.println("Tabla actualizada");
-                        return;
-                        // Éxito
-                    } else {
-                        // Sería conveniente lanzar una excepción en este caso y devolver un código de error apropiado
+                    if (rowsAffected < 1) {
+                        throw new HttpExceptionCode(500);
                     }
                 } else {
-                    // Sería conveniente lanzar una excepción en este caso y devolver un código de error apropiado
+                    throw new HttpExceptionCode(401);
                 }
+            } else {
+                throw new HttpExceptionCode(404);
             }
-            // Sería conveniente lanzar una excepción en este caso y devolver un código de error apropiado
         } catch (SQLException throwables) {
             throwables.printStackTrace();
         }
-        // Sería conveniente lanzar una excepción en este caso y devolver un código de error apropiado
     }
 
 }
